requirements.txt
asgiref==3.8.1
cachetools==5.5.0
cloudinary==1.41.0
crispy-bootstrap5==0.7
dj-database-url==0.5.0
Django==4.2.16
django-allauth==0.57.2
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
django-summernote==0.8.20.0
gunicorn==20.1.0
oauthlib==3.2.2
orjson==3.10.11
psycopg2==2.9.10
PyJWT==2.9.0
python3-openid==3.2.0
requests-oauthlib==2.0.0
shortuuid==1.0.13
sqlparse==0.5.1
whitenoise==6.8.2
